******************************************* ***** database need to be destroyed ! ***** *******************************************
* database : acquisition and acquisitionSummary : remove id column (not need and avoid declare an index, so data size is now half)
* database : acquisitionSummary : computations are done for each data insert (may change in future)
* plugin : RfxLan : SensorBasic and ControlBasic (in progress, works with our equipments)
* plugin : OpenZWave : fix mismatch type warnings

+++ trunk/sources/plugins/ZWave/OpenZWaveNode.cpp
@@ -10,7 +10,7 @@ COpenZWaveNode::COpenZWaveNode()
    throw shared::exception::CException("Should not happen");
 }
 
-COpenZWaveNode::COpenZWaveNode(const int homeId, const int nodeId)
+COpenZWaveNode::COpenZWaveNode(const uint32 homeId, const uint8 nodeId)
    :m_homeId(homeId), m_nodeId(nodeId)
 {
 }
@@ -57,7 +57,7 @@ const shared::plugin::yadomsApi::commands::IHistorizable & COpenZWaveNode::getLa
 }
 
 
-const bool COpenZWaveNode::match(const int homeId, const int nodeId)
+const bool COpenZWaveNode::match(const uint32 homeId, const uint8 nodeId)
 {
    return m_homeId == homeId && m_nodeId == nodeId;
 }
